+++ src/api/c++/src/cursor/resultsetbuffersize.C
@@ -4,17 +4,17 @@
 #include <config.h>
 #include <sqlrelay/sqlrclient.h>
 
-void sqlrcursor::setResultSetBufferSize(int rows) {
+void sqlrcursor::setResultSetBufferSize(uint32_t rows) {
 	rsbuffersize=rows;
 	if (sqlrc->debug) {
 		sqlrc->debugPreStart();
 		sqlrc->debugPrint("Result Set Buffer Size: ");
-		sqlrc->debugPrint((long)rows);
+		sqlrc->debugPrint((int32_t)rows);
 		sqlrc->debugPrint("\n");
 		sqlrc->debugPreEnd();
 	}
 }
 
-int sqlrcursor::getResultSetBufferSize() {
+uint32_t sqlrcursor::getResultSetBufferSize() {
 	return rsbuffersize;
 }
