+++ src/api/mysql/mysql.C
@@ -785,7 +785,7 @@ my_bool mysql_bind_param(MYSQL_STMT *stmt, MYSQL_BIND *bind) {
 	unsigned long	paramcount=mysql_param_count(stmt);
 	for (unsigned long i=0; i<paramcount; i++) {
 
-		char		*variable=charstring::parseNumber(i);
+		char		*variable=charstring::parseNumber((uint32_t)i);
 		sqlrcursor	*cursor=stmt->result->sqlrcur;
 		switch (bind[i].buffer_type) {
 			case MYSQL_TYPE_NULL: {
